+++ packages/create-farm/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-farm
 
+## 1.2.17
+
+### Patch Changes
+
+- bump @farmfe/core to 0.13.7
+
+## 1.2.16
+
+### Patch Changes
+
+- preserve import.meta when targetEnv is Node
+
 ## 1.2.15
 
 ### Patch Changes